requirements.txt
asgiref==3.8.1
asttokens==2.4.1
coverage==7.6.0
decorator==5.1.1
Django==5.0.6
django-filter==24.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
executing==2.0.1
ipython==8.25.0
jedi==0.19.1
matplotlib-inline==0.1.7
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
prompt_toolkit==3.0.47
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
PyJWT==2.8.0
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
traitlets==5.14.3
wcwidth==0.2.13
